Factor out geotemporal tools
